Keep the protocol era as typed state on the HTTP transport
