+++ src/Client/Transport/HttpTransport.php
@@ -13,11 +13,11 @@ use Laravel\Mcp\Client\Contracts\Transport;
 use Laravel\Mcp\Client\Exceptions\AuthorizationRequiredException;
 use Laravel\Mcp\Client\Exceptions\RequestRejectedException;
 use Laravel\Mcp\Client\OAuth\WwwAuthenticateChallenge;
-use Laravel\Mcp\Client\Transport\Eras\Era;
-use Laravel\Mcp\Client\Transport\Eras\LegacyEra;
 use Laravel\Mcp\Enums\ProtocolVersion;
+use Laravel\Mcp\Enums\RequestHeader;
 use Laravel\Mcp\Exceptions\ClientException;
 use Laravel\Mcp\Exceptions\SessionExpiredException;
+use Laravel\Mcp\Transport\JsonRpcRequest;
 use Psr\Http\Message\StreamInterface;
 use SensitiveParameter;
 use Throwable;
@@ -29,7 +29,11 @@ class HttpTransport implements Transport
     /** @var string|(Closure(): string)|null */
     protected string|Closure|null $token = null;
 
-    protected Era $era;
+    protected ?ProtocolVersion $protocolVersion = null;
+
+    protected ?string $sessionId = null;
+
+    protected bool $initialized = false;
 
     protected float $timeoutSeconds = 30.0;
 
@@ -41,7 +45,7 @@ class HttpTransport implements Transport
 
     public function __construct(protected string $url)
     {
-        $this->era = new LegacyEra;
+        //
     }
 
     public function connect(): void
@@ -101,7 +105,7 @@ class HttpTransport implements Transport
         $body = json_decode($message, true);
         $body = is_array($body) ? $body : [];
 
-        $hadSession = $this->era->hasSession();
+        $hadSession = $this->sessionId !== null;
 
         try {
             $response = Http::withHeaders($this->headers($body))
@@ -113,7 +117,7 @@ class HttpTransport implements Transport
             $this->failWith("HTTP request to [{$this->url}] failed: {$connectionException->getMessage()}");
         }
 
-        $this->era->inspect($response);
+        $this->captureSessionId($response);
 
         if ($response->status() === 401 || $response->status() === 403) {
             $challenge = WwwAuthenticateChallenge::parse($response->header('WWW-Authenticate'));
@@ -150,6 +154,8 @@ class HttpTransport implements Transport
             $this->failWith("Unexpected HTTP status [{$response->status()}] from endpoint [{$this->url}].");
         }
 
+        $this->initialized = true;
+
         if (str_contains($response->header('Content-Type'), 'text/event-stream')) {
             $this->readSseStream($response);
 
@@ -167,7 +173,14 @@ class HttpTransport implements Transport
 
     public function setProtocolVersion(ProtocolVersion $version): void
     {
-        $this->era = $this->era->speaking($version);
+        $previous = $this->protocolVersion;
+
+        if ($previous instanceof ProtocolVersion && $previous->usesDiscovery() !== $version->usesDiscovery()) {
+            $this->sessionId = null;
+            $this->initialized = false;
+        }
+
+        $this->protocolVersion = $version;
     }
 
     public function receive(): string
@@ -201,7 +214,7 @@ class HttpTransport implements Transport
     {
         $headers = [
             'Accept' => 'application/json, text/event-stream',
-            ...$this->era->headers($body),
+            ...$this->eraHeaders($body),
         ];
 
         $token = $this->token instanceof Closure ? (string) ($this->token)() : $this->token;
@@ -221,6 +234,60 @@ class HttpTransport implements Transport
         }
 
         return $headers;
+    }
+
+    /**
+     * @param  array<string, mixed>  $body
+     * @return array<string, string>
+     */
+    protected function eraHeaders(array $body): array
+    {
+        $version = $this->protocolVersion;
+
+        if ($version instanceof ProtocolVersion && $version->usesDiscovery()) {
+            return [
+                RequestHeader::PROTOCOL_VERSION->value => $version->value,
+                ...$this->mirroredHeaders($body),
+            ];
+        }
+
+        $headers = [];
+
+        if ($this->sessionId !== null) {
+            $headers['MCP-Session-Id'] = $this->sessionId;
+        }
+
+        if ($this->initialized && ($body['method'] ?? null) !== 'initialize') {
+            $headers[RequestHeader::PROTOCOL_VERSION->value] = ($version ?? ProtocolVersion::V2025_11_25)->value;
+        }
+
+        return $headers;
+    }
+
+    /**
+     * @param  array<string, mixed>  $body
+     * @return array<string, string>
+     */
+    protected function mirroredHeaders(array $body): array
+    {
+        if (! isset($body['id']) || ! is_string($body['method'] ?? null)) {
+            return [];
+        }
+
+        return (new JsonRpcRequest(
+            id: is_int($body['id']) || is_string($body['id']) ? $body['id'] : 0,
+            method: $body['method'],
+            params: is_array($body['params'] ?? null) ? $body['params'] : [],
+        ))->mirroredHeaders();
+    }
+
+    protected function captureSessionId(ClientResponse $response): void
+    {
+        $sessionId = $response->header('MCP-Session-Id');
+
+        if ($sessionId !== '') {
+            $this->sessionId = $sessionId;
+        }
     }
 
     protected function readSseStream(ClientResponse $response): void
@@ -274,7 +341,7 @@ class HttpTransport implements Transport
 
     protected function terminateSession(): void
     {
-        if (! $this->era->hasSession()) {
+        if ($this->sessionId === null) {
             return;
         }
 
@@ -289,7 +356,9 @@ class HttpTransport implements Transport
 
     protected function reset(): void
     {
-        $this->era = new LegacyEra;
+        $this->protocolVersion = null;
+        $this->sessionId = null;
+        $this->initialized = false;
         $this->queue = [];
     }
